RSpec: move the definition of MyRecord
