+++ spec/spec_helper.rb
@@ -1,9 +1,6 @@
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
 require 'active_record'
 require 'active_record_upsert'
-
-require 'setup'
-
 require 'database_cleaner'
 
 RSpec.configure do |config|
@@ -21,6 +18,29 @@ RSpec.configure do |config|
       end
       ActiveRecord::Base.connection.add_index :my_records, :wisdom, unique: true
     end
+
+    class MyRecord < ActiveRecord::Base
+      before_save :before_s
+      after_save :after_s
+      before_create :before_c
+      after_create :after_c
+      after_commit :after_com
+
+      def before_s
+      end
+
+      def after_s
+      end
+
+      def before_c
+      end
+
+      def after_c
+      end
+
+      def after_com
+      end
+    end    
 
     DatabaseCleaner.strategy = :transaction
     DatabaseCleaner.clean_with(:truncation)
